HUT is the Hid Usage Table crate
This is largely generated but also static set of enums and a few helpers
around it.

This is split out the hut module from the hideport crate.

This is effectively independent anyway and shouldn't ever need to be
updated (or updated much), so let's move this into its own crate.