Implement reserved pages
These fell through the cracks as general usage pages before and we do
have some devices that use 0xff (which hidtools interprets as vendor
page but the HUT standard says is in the reserved range).

So let's add the same combo of ReservedUsagePage and ReservedPage that
we have for vendor pages already.