Add support for no_std
This apparently requires dropping thiserror which isn't no_std
compatible but it's easy enough to implement the few errors we have.

The only thing really disabled in the no_std implementation is the
name() method (and thus Display) on the various enums since that
requires a string allocation but that shouldn't really be a problem for
those actually needing no_std.

Closes #16